src/requirements/ecs.txt
Flask==1.0.2
flask-restx==0.2.0
boto3==1.9.123
gunicorn==20.1.0
gevent==21.8.0
flask-cors==3.0.7
python-jose[cryptography]==3.3.0
cachecontrol[filecache]==0.12.6
